Fix lint and move helper to utils file
